Update incorrect commands
There were what I suspect are typos in the commands for run_picam.sh and the example script with redirection of standard error and output. 2&>1 should be 2>&1. (and a missing space)
The nginx start command does not work since nginx will already be running after install, a restart is required for the changes to take effect.